fix based on review

+++ backend/app/core/config.py
@@ -25,7 +25,7 @@ class Settings(BaseSettings):
     ]
 
     # Delay between Jina AI Reader fetches to respect rate limits
-    JINA_FETCH_DELAY_SECONDS: int = 4
+    JINA_FETCH_DELAY_SECONDS: float = 4.0
 
     # Example of another setting that might be needed later
     # ANTHROPIC_API_KEY: Optional[str] = None
